🎨 Refactor frontend in components
* Move Product out of App.
* Split out AssetsDisplay component.
* Split out KpisDisplay component.
* Split out Chart component.
