+++ frontend/src/dateUtils.ts
@@ -1,0 +1,26 @@
+export const getTodayMidnight = () => {
+  const today = new Date();
+  today.setUTCHours(0, 0, 0, 0);
+  return today;
+};
+
+// https://stackoverflow.com/a/54452875/931156
+export function deltaDate(
+  input: Date,
+  years: number,
+  months: number,
+  days: number
+) {
+  return new Date(
+    input.getFullYear() + years,
+    input.getMonth() + months,
+    Math.min(
+      input.getDate() + days,
+      new Date(
+        input.getFullYear() + years,
+        input.getMonth() + months + 1,
+        0
+      ).getDate()
+    )
+  );
+}
