+++ frontend/src/simulationContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+import { SimulationContext } from "./types";
+
+export const SimulationReactContext = createContext<SimulationContext>({
+  address: "",
+  startDate: new Date(),
+  today: new Date(),
+});
